% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmcPrior.R
\docType{package}
\name{ggdmcPrior}
\alias{ggdmcPrior-package}
\alias{ggdmcPrior}
\title{ggdmc module for standard and truncated distributions}
\description{
Provides tools for specifying and evaluating standard and truncated
probability distributions, with support for log-space computation and
joint distribution specification. It enables Bayesian computation for
cognition models and includes utilities for density calculation, sampling,
and visualisation, facilitating prior distribution specification and model
assessment in hierarchical Bayesian frameworks.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com>
}
\keyword{internal}
\keyword{package}
