% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{prior-class}
\alias{prior-class}
\title{An S4 class  to represent the joint distribution.}
\description{
An S4 class  to represent the joint distribution.
}
\section{Slots}{

\describe{
\item{\code{nparameter}}{the number of free parameters}

\item{\code{pnames}}{the names of free parameters}

\item{\code{p_prior}}{a list storing the subject-level prior (joint) distribution.}

\item{\code{h_prior}}{a list storing the population-level prior (joint) distribution
(location and scale distributions). `h` stands for hyper-level.}
}}

