% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tnorm}
\alias{tnorm}
\alias{rtnorm}
\alias{ptnorm}
\alias{dtnorm}
\title{Truncated normal distribution}
\usage{
rtnorm(n, p0, p1, lower, upper)

ptnorm(x, p0, p1, lower, upper, lower_tail, log_p = FALSE)

dtnorm(x, p0, p1, lower, upper, log_p = FALSE)
}
\arguments{
\item{n}{Number of random variates to generate (for `rtnorm`).}

\item{p0}{Mean of the underlying normal distribution.}

\item{p1}{Standard deviation of the underlying normal distribution.}

\item{lower}{Lower truncation bound.}

\item{upper}{Upper truncation bound.}

\item{x}{Vector of quantiles (for `dtnorm` and `ptnorm`).}

\item{lower_tail}{Logical; if TRUE (default), probabilities are
\deqn{P[X \leq x]}, otherwise P[X > x].}

\item{log_p}{Logical; if TRUE, probabilities/densities p are returned as
log(p).}
}
\value{
\describe{
  \item{`rtnorm`}{NumericVector of random variates from the truncated
         normal}
 \item{`ptnorm`}{NumericVector of cumulative probabilities}
 \item{`dtnorm`}{NumericVector of density values}
}

 @details
 These functions implement the truncated normal distribution:
 \itemize{
   \item `rtnorm`: Random number generation using inverse transform sampling
   \item `ptnorm`: Cumulative distribution function (CDF)
   \item `dtnorm`: Probability density function (PDF)
 }

The distribution is defined as:
\deqn{X \sim \mathcal{N}(\mu, \sigma^2), \text{truncated to } [a, b]}
where \eqn{\mu = p0}, \eqn{\sigma = p1}, \eqn{a = lower}, \eqn{b = upper}.
}
\description{
Density, distribution function, and random generation for the truncated
normal distribution with mean `p0`, standard deviation `p1`, and bounds
[`lower`, `upper`].
}
\examples{
n <- 1e5
p0 <- 0
p1 <- 1
lower <- 0
upper <- Inf
rtnorm_dat <- rtnorm(n, p0, p1, lower, upper)

## dtnorm example
x <- seq(-5, 5, length.out = 1e3)
dtnorm_dat <- dtnorm(x, p0, p1, lower = -2, upper = 2, log_p = FALSE)
q <- seq(-5, 5, length.out = 1e3)

ptnorm_dat <- ptnorm(q, p0, p1, lower = -2, upper = 3, lower_tail = TRUE,
                 log_p = FALSE)

cex_lab <- 1
cex_axis <- 0.5
line_width <- 1.5

hist(rtnorm_dat, breaks = "fd", freq = FALSE, xlab = "", cex.lab = cex_lab,
cex.axis = cex_axis, main = "" )
plot(x, dtnorm_dat, type = "l", lwd = line_width, xlab = "",
ylab = "Density", cex.lab = cex_lab, cex.axis = cex_axis, main = "")

}
\references{
\itemize{
  \item RcppTN package by Jonathan Olmsted (RcppTN 0.1-8)
        \url{https://github.com/olmjo/RcppTN}
  \item msm package by Christopher Jackson
        \url{https://cran.r-project.org/package=msm}
  \item Robert, C.P. (1995). "Simulation of truncated normal variables".
    Statistics and Computing, 5(2), 121-125. \doi{10.1007/BF00143942}
 }
}
