% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dprior}
\alias{dprior}
\alias{rprior}
\title{Density and random number generation}
\usage{
dprior(p_prior_r, parameters_r)

rprior(p_prior_r, n = 1L)
}
\arguments{
\item{p_prior_r}{A List containing distribution specifications (usually
specified in \code{BuildPrior} function).}

\item{parameters_r}{For `dprior`: A NumericVector of parameter values at
which to evaluate the prior density.}

\item{n}{For `rprior`: Number of random samples to generate (default = 1).}
}
\value{
\describe{
  \item{`dprior`}{an R vector of prior density values}
  \item{`rprior`}{an R matrix of random samples (n rows × nparameter
              columns)}
 }
}
\description{
\code{dprior} computes the density for a joint distribution. \code{rprior}
generates random numbers based on the specification in the \code{p_prior_r}
}
\details{
These functions provide the core computational methods for prior
distributions:
 \itemize{
  \item `dprior`: Evaluates the prior probability density function
  \item `rprior`: Generates random samples from the prior distribution
}
}
\examples{
p0 <- c(A = 0.15, B = 0.45, mean_v = 2.25, sd_v = 0.15, t0 = 0.2)
p1 <- rep(0.1, 5)
p_prior <- BuildPrior(
    p0 = p0,
    p1 = p1,
    lower = rep(NA, 5),
    upper = rep(NA, 5),
    dist = rep("tnorm", 5),
    log_p = rep(TRUE, 5)
)

# Test dprior -----------------
parameters <- seq(0.1, 0.5, 0.1)
res0 <- dprior(p_prior, parameters)
res1 <- dnorm(parameters, p0, 0.1, TRUE)
print(res0)
print(res1)

# Test rprior -----------------
res2 <- rprior(p_prior, 1)
print(res2)
res3 <- rprior(p_prior, 2)
print(res3)
}
