% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmcPrior.R
\docType{package}
\name{ggdmcPrior}
\alias{ggdmcPrior-package}
\alias{ggdmcPrior}
\title{ggdmc module for standard and truncated distributions}
\description{
\pkg{ggdmcPrior} provides functions for specifying and evaluating standard
distributions, designed to work with the ggdmc package. It supports Bayesian
computation and includes utilities for density calculation, sampling,
and visualisation.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com>
}
\keyword{internal}
\keyword{package}
