% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{prior-class}
\alias{prior-class}
\title{An S4 class  to represent the joint distribution.}
\description{
An S4 class  to represent the joint distribution.
}
\section{Slots}{

\describe{
\item{\code{nparameter}}{the number of parameters}

\item{\code{pnames}}{the names of parameters}

\item{\code{p_prior}}{a prior object storing the subject-level prior distribution.}

\item{\code{h_prior}}{a prior object storing the population-level distribution
(location and scale distributions).}
}}

